Worked around crashes when plugin is loaded as an external plugin.
The plugin used to crash when xfce4-panel was compiled with
--enable-debug and the plugin was configured as an external module
(X-XFCE-Internal=FALSE in indicator.desktop file).

This was caused by the wrapper binary setting a more restrictive
fatal mask than the xfce4-panel binary. As we are linking external
code (indicators) we have no control over their use of g_warning or
g_critical.

This commit relaxes the fatal mask so that the use of g_warning or
g_critical does not crash the plugin.