Fixed #9040.
This required getting replacing XfcePanelImage with GtkImage. The former
does not work reliably with non-square icons and added automatization
is not needed in this case.

Only horizontal non-square icons are supported (vertical will be squeezed
into a square shape). Non-square icons are rendered in a single row
(like indicators with labels).